add: 굿즈 찜 api 추가
